requirements.txt
annotated-types==0.7.0
APScheduler==3.10.4
blinker==1.8.2
boto3==1.35.3
botocore==1.35.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
environs==11.0.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Cors==4.0.1
flask-openapi3==3.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
importlib_metadata==8.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
SQLAlchemy==2.0.32
tomli==2.0.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==1.26.19
websockets==12.0
Werkzeug==3.0.4
zipp==3.20.0
